+++ packages/gov-ui-kit/src/components/input/inputDate/index.ts
@@ -1,0 +1,1 @@
+export { InputDate, type IInputDateProps } from './inputDate';