+++ packages/gov-ui-kit/src/components/input/inputDate/inputDate.stories.tsx
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { InputDate } from './inputDate';
+
+const meta: Meta<typeof InputDate> = {
+    title: 'components/Input/InputDate',
+    component: InputDate,
+    tags: ['autodocs'],
+    parameters: {
+        design: {
+            type: 'figma',
+            url: 'https://www.figma.com/file/jfKRr1V9evJUp1uBeyP3Zz/v1.0.0?type=design&node-id=10080-1466&mode=design&t=2bLCEeKZ7ueBboTs-4',
+        },
+    },
+};
+
+type Story = StoryObj<typeof InputDate>;
+
+/**
+ * Default usage example of the InputDate component.
+ */
+export const Default: Story = {
+    args: {},
+};
+
+export default meta;