+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 -   Minimum `tailwindcss` version required
+-   Fix disabled input style on Firefox
+
+### Added
+
+-   Implement `InputDate` component
 
 ## [1.0.8] - 2024-01-17